+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Order item finder downdrop by name and quantity - issue #118
 - Improved input validation
+- Auto scaling label title
 
 ### Fixed
 - Do not wrap button fields - issue #122